v0.4.2
 - Broken, casting error with Packet, need to figure out solution
 - Made a Packet class, maybe we need to delete it

+++ Project REWIND/src/sprites/player/Player.java
@@ -37,7 +37,7 @@ public class Player extends Sprite {
 	
 	private Point2D.Double bulletPoint;
 	
-	public Player(PImage img, int x, int y) {
+	public Player(PImage img, double x, double y) {
 		super(img, x, y, PLAYER_WIDTH, PLAYER_HEIGHT);
 		playerType = 1;
 		health = 5;
